RDCC-3226: refactoring unit tests
