+++ src/main/java/uk/gov/hmcts/reform/cwrdapi/service/CaseWorkerService.java
@@ -5,7 +5,6 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.cwrdapi.client.domain.ServiceRoleMapping;
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.CaseWorkersProfileCreationRequest;
-import uk.gov.hmcts.reform.cwrdapi.controllers.response.CaseWorkerProfilesDeletionResponse;
 import uk.gov.hmcts.reform.cwrdapi.controllers.response.IdamRolesMappingResponse;
 import uk.gov.hmcts.reform.cwrdapi.domain.CaseWorkerProfile;
 
@@ -39,22 +38,6 @@ public interface CaseWorkerService {
      * @return CaseWorkerProfile
      */
     ResponseEntity<Object> fetchCaseworkersById(List<String> caseWorkerIds);
-
-    /**
-     * Deletes a caseworker by their ID.
-     *
-     * @param userId String
-     * @return CaseWorkerProfilesDeletionResponse
-     */
-    CaseWorkerProfilesDeletionResponse deleteByUserId(String userId);
-
-    /**
-     * Deletes caseworkers by associated email pattern.
-     *
-     * @param emailPattern String
-     * @return CaseWorkerProfilesDeletionResponse
-     */
-    CaseWorkerProfilesDeletionResponse deleteByEmailPattern(String emailPattern);
 
     /**
      * Returns the staff details for Refresh Assignments.
