+++ src/main/java/uk/gov/hmcts/reform/cwrdapi/service/impl/CaseWorkerServiceImpl.java
@@ -13,7 +13,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import uk.gov.hmcts.reform.cwrdapi.client.domain.Location;
 import uk.gov.hmcts.reform.cwrdapi.client.domain.Role;
 import uk.gov.hmcts.reform.cwrdapi.client.domain.ServiceRoleMapping;
@@ -33,7 +32,6 @@ import uk.gov.hmcts.reform.cwrdapi.controllers.request.LanguagePreference;
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.UserCategory;
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.UserProfileCreationRequest;
 import uk.gov.hmcts.reform.cwrdapi.controllers.request.UserTypeRequest;
-import uk.gov.hmcts.reform.cwrdapi.controllers.response.CaseWorkerProfilesDeletionResponse;
 import uk.gov.hmcts.reform.cwrdapi.controllers.response.IdamRolesMappingResponse;
 import uk.gov.hmcts.reform.cwrdapi.controllers.response.LrdOrgInfoServiceResponse;
 import uk.gov.hmcts.reform.cwrdapi.controllers.response.UserProfileCreationResponse;
@@ -76,7 +74,6 @@ import java.util.stream.Collectors;
 
 import static java.lang.String.format;
 import static java.lang.String.valueOf;
-import static java.util.Collections.singletonList;
 import static java.util.Objects.isNull;
 import static java.util.Objects.nonNull;
 import static java.util.Objects.requireNonNull;
@@ -87,8 +84,6 @@ import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 import static org.apache.commons.lang3.BooleanUtils.isNotTrue;
 import static org.apache.commons.lang3.BooleanUtils.isTrue;
 import static org.springframework.http.HttpStatus.CONFLICT;
-import static org.springframework.http.HttpStatus.NOT_FOUND;
-import static org.springframework.http.HttpStatus.NO_CONTENT;
 import static org.springframework.util.CollectionUtils.isEmpty;
 import static uk.gov.hmcts.reform.cwrdapi.util.CaseWorkerConstants.ALREADY_SUSPENDED_ERROR_MESSAGE;
 import static uk.gov.hmcts.reform.cwrdapi.util.CaseWorkerConstants.IDAM_STATUS;
@@ -703,73 +698,6 @@ public class CaseWorkerServiceImpl implements CaseWorkerService {
             }
         }
     }
-
-    @Transactional
-    public CaseWorkerProfilesDeletionResponse deleteByUserId(String userId) {
-        CaseWorkerProfilesDeletionResponse deletionResponse = new CaseWorkerProfilesDeletionResponse();
-
-        Response userProfileResponse = userProfileFeignClient.deleteUserProfile(userId, null);
-
-        if (userProfileResponse.status() == NO_CONTENT.value() || userProfileResponse.status() == NOT_FOUND.value()) {
-            Optional<CaseWorkerProfile> caseWorkerProfile = caseWorkerProfileRepo.findByCaseWorkerId(userId.trim());
-
-            return validateUserAfterUpDelete(caseWorkerProfile, userId, userProfileResponse.status());
-
-        } else {
-            deletionResponse.setMessage("UP Delete request failed for userId: " + userId
-                    + ". With the following UP message: " + userProfileResponse.reason());
-            deletionResponse.setStatusCode(userProfileResponse.status());
-            return deletionResponse;
-        }
-    }
-
-    @Transactional
-    public CaseWorkerProfilesDeletionResponse deleteByEmailPattern(String emailPattern) {
-        CaseWorkerProfilesDeletionResponse deletionResponse = new CaseWorkerProfilesDeletionResponse();
-
-        List<CaseWorkerProfile> userProfiles = caseWorkerProfileRepo.findByEmailIdIgnoreCaseContaining(emailPattern);
-
-        if (userProfiles.isEmpty()) {
-            throw new ResourceNotFoundException("No User Profiles found for email pattern: " + emailPattern);
-        }
-
-        Response userProfileResponse = userProfileFeignClient.deleteUserProfile(null, emailPattern);
-
-        if (userProfileResponse.status() == NO_CONTENT.value() || userProfileResponse.status() == NOT_FOUND.value()) {
-            return deleteUserProfiles(userProfiles);
-
-        } else {
-            deletionResponse.setMessage("UP Delete request failed for emailPattern: " + emailPattern);
-            deletionResponse.setStatusCode(userProfileResponse.status());
-            return deletionResponse;
-        }
-    }
-
-    public CaseWorkerProfilesDeletionResponse validateUserAfterUpDelete(
-            Optional<CaseWorkerProfile> caseWorkerProfile, String userId, int status) {
-        CaseWorkerProfilesDeletionResponse deletionResponse = new CaseWorkerProfilesDeletionResponse();
-
-        if (caseWorkerProfile.isPresent()) {
-            return deleteUserProfiles(singletonList(caseWorkerProfile.get()));
-
-        } else if (status == NO_CONTENT.value()) {
-            deletionResponse.setMessage("User deleted in UP but was not present in CRD with userId: " + userId);
-
-        } else {
-            deletionResponse.setMessage("User was not present in UP or CRD with userId: " + userId);
-        }
-
-        deletionResponse.setStatusCode(status);
-        return deletionResponse;
-    }
-
-    public CaseWorkerProfilesDeletionResponse deleteUserProfiles(List<CaseWorkerProfile> userProfiles) {
-        caseWorkerProfileRepo.deleteAll(userProfiles);
-
-        return new CaseWorkerProfilesDeletionResponse(NO_CONTENT.value(),
-                "Case Worker Profiles successfully deleted");
-    }
-
 
     public boolean isUserSuspended(UserProfileUpdatedData userProfileUpdatedData, String userId,
                                    String origin, long rowId) {
